pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Alexander Gedz <gedz.buh@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
black = "^24.8.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.1.1"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
pytz = "^2024.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
setuptools = "^75.1.0"
coverage = "^7.6.2"
python-dotenv = "^1.0.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
